refactor(virrun): file-organization sweep over the exec util services
The probe cache's options bag is a model. The private helpers inside
the hidden exec, the worktree reader and the process start-time probe
each become a service of their own: the stderr encoding sniff and the
error's stderr read, the git common-directory resolution, and the
Linux and Windows halves of the start-time read.
