+++ .agents/ledgers/file-organization.md
@@ -42,7 +42,7 @@ The question is where a thing lives and whether it exists twice — one export p
 | `packages/azure`, `apps/functions`, `packages/azure-mock`                       | 2026-09-12 | cross-package placement: an Azure helper two packages need lives in `db`                        |
 | `packages/virrun` — `services/exec/wsl`                                         | 2026-09-12 |                                                                                                 |
 | `packages/virrun` — `services/exec/snapshot`                                    | 2026-09-12 |                                                                                                 |
-| `packages/virrun` — `services/exec/util`                                        | —          |                                                                                                 |
+| `packages/virrun` — `services/exec/util`                                        | 2026-09-12 |                                                                                                 |
 | `packages/virrun` — `services/exec` — the rest                                  | —          | `cache`, `os`, `bwrap`, `vfs`, `differential`, `store`, `native`, `test`                        |
 | `packages/virrun` — `services` — the rest                                       | —          | `cli`, `configuration`, `source`, `vfs`, `virrun`                                               |
 | `packages/virrun` — `models`, root                                              | —          |                                                                                                 |
