+++ packages/virrun/src/models/exec/util/ProbeCacheOptions.ts
@@ -1,0 +1,8 @@
+import type { KeyedCache } from "#src/models/exec/KeyedCache";
+
+export interface ProbeCacheOptions<TValue> {
+  probe: () => TValue;
+  readPersistedCache: (key: string) => TValue | undefined;
+  shouldPersist: (value: TValue) => boolean;
+  writePersistedCache: (cache: Pick<KeyedCache<TValue>, "key" | "value">) => void;
+}
